+++ http/src/main/scala/com/twitter/finatra/http/marshalling/mapper.scala
@@ -12,7 +12,7 @@ object mapper {
         if (length == 0) {
           throw new UnsupportedOperationException(
             "Injecting request attributes (e.g. QueryParam, Header, etc) not supported when explicitly calling " +
-              "FinatraObjectMapper.parse. Instead use a 'case class' input parameter on a Controller callback " +
+              "ScalaObjectMapper.parse. Instead use a 'case class' input parameter on a Controller callback " +
               "(e.g. get('/') { r: ClassWithRequestAttributes => ... } ).")
         }
       }